requirements.txt
# AI助手后端依赖

# Web框架
flask>=2.3.0
flask-cors>=4.0.0

# HTTP请求
requests>=2.31.0
urllib3>=2.0.0

# Redis缓存
redis>=5.0.0

# 环境变量
python-dotenv>=1.0.0

# 其他工具
gunicorn>=21.2.0
